// ui/src/main/angular/ui/modules/lipidcompass-backend-client/model/databaseStatistics.ts
/**
 * LipidCompass Backend API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DatabaseStatistics { 
    releaseVersion?: string;
    releaseDate?: string;
    categories?: number;
    classes?: number;
    species?: number;
    molecularSpecies?: number;
    snPosition?: number;
    structureDefined?: number;
    completeStructure?: number;
    fullStructure?: number;
    noLevel?: number;
    undefinedLevel?: number;
    experiments?: number;
    results?: number;
    crossReferences?: number;
    taxonomicSpecies?: number;
    tissues?: number;
    diseases?: number;
    cellTypes?: number;
    lipidQuantities?: number;
    studyVariableFactors?: number;
}